newly created files for use in recording appear in a .stubs folder, and are moved out of it when recording stops
git-svn-id: svn://localhost/ardour2/branches/3.0@7426 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ libs/ardour/ardour/types.h
@@ -27,6 +27,7 @@
 #include <istream>
 #include <vector>
 #include <boost/shared_ptr.hpp>
+#include <sys/types.h>
 
 #include <inttypes.h>
 #include <jack/types.h>
@@ -440,7 +441,7 @@ namespace ARDOUR {
 
 	struct CleanupReport {
 		std::vector<std::string> paths;
-		int64_t                  space;
+		size_t                   space;
 	};
 
 	enum PositionLockStyle {
